Renamed class from PosxmlVersion to ApplicationVersion.
